pip_requirements.txt
black
pylint
pytest
pytest-cov
slackclient==2.2.1 
